#527 Updated project staging profle creation to service
